requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.5
django-ckeditor==5.4.0
django-js-asset==1.1.0
django-recaptcha==1.4.0
gunicorn==19.8.1
Pillow==5.1.0
pytz==2018.4
static3==0.7.0
psycopg2>=2.6.1
